[perf] 일정 계획 페이지 수정
